@@ -20,20 +20,24 @@
     "@notionhq/client": "^2.2.14",
     "@prisma/client": "^5.13.0",
     "@radix-ui/react-dialog": "^1.0.5",
-    "@radix-ui/react-popover": "^1.0.2",
+    "@radix-ui/react-popover": "^1.0.7",
+    "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-tooltip": "^1.0.2",
     "@sendgrid/client": "^8.0.0",
     "@sendgrid/mail": "^8.0.0",
     "@stripe/react-stripe-js": "^2.4.0",
     "@stripe/stripe-js": "^2.3.0",
-    "@techwithty/cyberoni-schema": "^1.0.13",
+    "@techwithty/cyberoni-schema": "^1.0.14",
     "@vercel/analytics": "^0.1.11",
     "@vercel/og": "^0.6.2",
     "@vercel/speed-insights": "^1.0.6",
     "bcrypt": "^5.1.1",
+    "class-variance-authority": "^0.7.0",
     "classnames": "^2.5.1",
     "cloudinary": "^1.41.2",
+    "clsx": "^2.1.1",
     "cookies-next": "^4.1.0",
+    "embla-carousel-react": "^8.0.4",
     "encoding": "^0.1.13",
     "focus-trap-react": "^10.0.2",
     "framer-motion": "^10.17.9",
@@ -41,23 +45,28 @@
     "jsonwebtoken": "^9.0.2",
     "lucide-react": "^0.307.0",
     "ms": "^2.1.3",
-    "next": "^13.5.6",
+    "next": "^14.2.3",
     "next-auth": "4.24.5",
     "next-recaptcha-v3": "^1.3.0",
     "nodemailer": "^6.9.9",
+    "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-alice-carousel": "^2.8.0",
     "react-calendly": "^4.3.0",
     "react-dom": "18.2.0",
+    "react-lottie": "^1.2.4",
     "react-markdown": "^8.0.4",
     "react-password-checklist": "^1.5.0",
     "react-tailwindcss-datepicker": "^1.6.6",
     "react-wrap-balancer": "^0.5.0",
     "schema-dts": "^1.1.2",
     "seedrandom": "^3.0.5",
+    "server-only": "^0.0.1",
     "sharp": "^0.33.2",
     "slugify": "^1.6.6",
     "stripe": "^14.14.0",
+    "tailwind-merge": "^2.3.0",
+    "tailwindcss-animate": "^1.0.7",
     "use-debounce": "^9.0.4",
     "xss": "^1.0.14"
   },
@@ -74,6 +83,7 @@
     "@types/node": "^20.8.2",
     "@types/react": "^18.2.25",
     "@types/react-dom": "^18.2.10",
+    "@types/react-lottie": "^1.2.10",
     "@types/seedrandom": "^3.0.8",
     "@types/supertest": "^2.0.16",
     "autoprefixer": "^10.4.13",
