@@ -1,7 +1,7 @@
-/** @type {import('tailwindcss').Config} */
-const plugin = require("tailwindcss/plugin");
+import type { Config } from "tailwindcss"
+import plugin from "tailwindcss/plugin";
 
-module.exports = {
+const config = {
   content: ["./app/**/*.{js,ts,jsx,tsx}", "./components/**/*.{js,ts,jsx,tsx}"],
   future: {
     hoverOnlyWhenSupported: true,
@@ -14,6 +14,8 @@ module.exports = {
         default: ["var(--font-inter)", "system-ui", "sans-serif"],
         abel: ["var(--font-abel)", "system-ui", "sans", "serif"],
         nunito: ["var(--font-nunito)", "system-ui", "sans", "serif"],
+        kyiv: ["var(--font-kyiv)", "system-ui", "sans", "serif"],
+
       },
       animation: {
         // Fade up and down
@@ -46,38 +48,38 @@ module.exports = {
         // Fade up and down
         "fade-up": {
           "0%": {
-            opacity: 0,
+            opacity: "0",
             transform: "translateY(10px)",
           },
           "80%": {
-            opacity: 0.6,
+            opacity: "0.6",
           },
           "100%": {
-            opacity: 1,
+            opacity: "1",
             transform: "translateY(0px)",
           },
         },
         "fade-down": {
           "0%": {
-            opacity: 0,
+            opacity: "0",
             transform: "translateY(-10px)",
           },
           "80%": {
-            opacity: 0.6,
+            opacity: "0.6",
           },
           "100%": {
-            opacity: 1,
+            opacity: "1",
             transform: "translateY(0px)",
           },
         },
         // Tooltip
         "slide-up-fade": {
-          "0%": { opacity: 0, transform: "translateY(6px)" },
-          "100%": { opacity: 1, transform: "translateY(0)" },
+          "0%": { opacity: "0", transform: "translateY(6px)" },
+          "100%": { opacity: "1", transform: "translateY(0)" },
         },
         "slide-down-fade": {
-          "0%": { opacity: 0, transform: "translateY(-6px)" },
-          "100%": { opacity: 1, transform: "translateY(0)" },
+          "0%": { opacity:"0", transform: "translateY(-6px)" },
+          "100%": { opacity: "1", transform: "translateY(0)" },
         },
       },
       colors: {
@@ -98,4 +100,6 @@ module.exports = {
       
     }),
   ],
-};
+} satisfies Config
+
+export default config
