@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 
 export async function POST(req: NextRequest) {
-    console.log(await req.json());
+    //console.log(await req.json());
 
     return NextResponse.json({})
 }
