@@ -4,7 +4,7 @@ import React from 'react'
 
 function BlogLoader() {
     return (
-        <div className="w-full h-[60vh]">
+        <div className="w-full h-[60vh] overflow-hidden">
             <div className="">
                 <div className="w-full bg-white dark:bg-black py-5">
                     <div className="container mx-auto ">
@@ -20,7 +20,7 @@ function BlogLoader() {
                     </div>
                 </div>
 
-                <div className="container mx-auto my-10 flex flex-wrap overflow-hidden  py-10 px-10">
+                <div className="container mx-auto my-10 flex flex-wrap overflow-hidden  py-10 px-10 lg:w-[48rem]">
                     {new Array(30).fill(1).map((value, index) => {
 
                         return (
