@@ -5,13 +5,13 @@ import NavBar from "@/components/layout/navbar";
 import { Owner } from "@/data/ownerData";
 import { Analytics } from "@vercel/analytics/react";
 import cx from "classnames";
-import { Metadata } from "next";
+import { Metadata, Viewport } from "next";
 import { getServerSession } from "next-auth";
 import Head from "next/head";
 import { cookies } from "next/headers";
 import Script from "next/script";
 import { Suspense } from "react";
-import { abel, inter, nunito, sfPro } from "./fonts";
+import { abel, inter, kyiv, nunito, sfPro } from "./fonts";
 import "./globals.css";
 import { authOptions } from "@/lib/nextAuthAdapter";
 import { WebVitals } from "@/components/WebVitals";
@@ -40,18 +40,20 @@ export const metadata: Metadata = {
     locale: "en_US",
     type: "website",
   },
-  themeColor: "#0074e4",
   other: {
-    "fb:app_id": process.env.FACEBOOK_ID as string,   
+    "fb:app_id": process.env.FACEBOOK_ID as string,
   },
-  viewport: {
-    width: "device-width",
-    initialScale: 1,
-    maximumScale: 5
-  },
+
 
 };
 
+export const viewport: Viewport = {
+  width: "device-width",
+  initialScale: 1,
+  maximumScale: 5,
+  themeColor: "#0074e4",
+
+}
 export default async function RootLayout({
   children,
 }: {
@@ -92,6 +94,7 @@ export default async function RootLayout({
           inter.variable,
           abel.variable,
           nunito.variable,
+          kyiv.variable
         )} 
                         ${theme?.value} overflow-x-hidden antialiased`}
       >
