@@ -18,7 +18,7 @@ export type MegaMenuProps = Array<{
 
 
 
-function MegaMenu({ groups, setMenuOpen: closeMenu }: { groups: MegaMenuProps, setMenuOpen: Dispatch<SetStateAction<boolean>> }) {
+function MegaMenu({ groups,  closeMenu }: { groups: MegaMenuProps, closeMenu: ()=>void }) {
     return (
         <div className="shadow-lg dark:shadow-2xl rounded-lg bg-gray-50 dark:bg-slate-900 z-50 lg:px-5 lg:py-5 max-h-[65vh] xl:max-h-[85vh] overflow-y-auto scrollbar-thin scrollbar-thumb-gray-400  scrollbar-track-gray-50 scrollbar-thumb-rounded-md dark:scrollbar-track-slate-600 " >
             <div className="p-5 my-1">  
@@ -31,7 +31,7 @@ function MegaMenu({ groups, setMenuOpen: closeMenu }: { groups: MegaMenuProps, s
                                 {group.links.map((link, index) => (
                                     <div key={index} className="px-2 flex gap-2 w-1/2 my-2 items-center">
                                         {link.icon}
-                                        <Link onClick={() => closeMenu(false)} href={link.href} className="text-xl font-bold w-fit p-1">
+                                        <Link onClick={() => closeMenu()} href={link.href} className="text-xl font-bold w-fit p-1">
                                             {link.name}
                                             <div className="text-sm font-normal text-gray-600 dark:text-gray-200">{link.subTitle}</div>
                                         </Link>
