@@ -8,7 +8,7 @@ function ProgressBar({ container }: { container?: RefObject<HTMLElement> }) {
   const scrollY = scrolled(50);
 
   return (<motion.div
-    className={`fixed top-16 lg:top-24 left-0 h-2 w-screen bg-gradient-purple origin-left z-[100]`}
+    className={`fixed top-16 xl:top-24 left-0 h-2 w-screen bg-gradient-purple origin-left z-[100]`}
     style={{ scaleX: scrollYProgress }}
   />);
 }
