@@ -1,8 +1,9 @@
+import "server-only"
 import { PrismaClient } from "@prisma/client";
 import { connectOrCreateObject as connectTags } from "./tags";
 import { connectOrCreateObject as connectImages } from "./images";
 import { cleanHtmlString, getRandomFromArray } from "@/lib/utils";
-import { CommentDTO, CreateBlogDTO } from "./DTOs";
+import { CommentDTO, CreateBlogDTO, DisplayBlogDTO } from "./DTOs";
 import { getUserByEmail } from "./user";
 
 async function create(blog: CreateBlogDTO, prismaClient: PrismaClient) {
@@ -11,38 +12,40 @@ async function create(blog: CreateBlogDTO, prismaClient: PrismaClient) {
         data: {
             ...blog,
             images: await connectImages(blog.images, []),
-            tags:  connectTags(blog.tags, []) ,
-            author: { connect: { email: blog.author.email } }
-        }
+            tags: connectTags(blog.tags, []),
+            author: { connect: { email: blog.author.email } },
+        },
     });
-    return createdblog
-
-
+    return createdblog;
 }
 
-
-async function update(blogId: string, blog: CreateBlogDTO, prismaClient: PrismaClient) {
+async function update(
+    blogId: string,
+    blog: CreateBlogDTO,
+    prismaClient: PrismaClient,
+) {
     const blogs = prismaClient.blog;
-    const oldBlog = await blogs.findUnique({ where: { id: blogId }, include: { images: true, tags: true } })
+    const oldBlog = await blogs.findUnique({
+        where: { id: blogId },
+        include: { images: true, tags: true },
+    });
     const updatedBlog = await blogs.update({
         where: { id: blogId },
         data: {
             ...blog,
             images: await connectImages(blog.images, oldBlog!.images),
-            tags:  connectTags(blog.tags, oldBlog!.tags) ,
-            author: { connect: { email: blog.author.email } }
-        }
-    })
-    return updatedBlog
-
+            tags: connectTags(blog.tags, oldBlog!.tags),
+            author: { connect: { email: blog.author.email } },
+        },
+    });
+    return updatedBlog;
 }
-
 
 async function remove(blogId: string, prismaClient: PrismaClient) {
     const blogs = prismaClient.blog;
-    const existingblog = await blogs.findUnique({ where: { id: blogId } })
+    const existingblog = await blogs.findUnique({ where: { id: blogId } });
     if (existingblog) {
-        await blogs.delete({ where: { id: blogId } })
+        await blogs.delete({ where: { id: blogId } });
     }
 }
 async function read(blogId: string, prismaClient: PrismaClient) {
@@ -52,8 +55,8 @@ async function read(blogId: string, prismaClient: PrismaClient) {
             id: blogId,
 
             publishDate: {
-                lte: new Date()
-            }
+                lte: new Date(),
+            },
         },
         select: {
             userId: false,
@@ -68,46 +71,48 @@ async function read(blogId: string, prismaClient: PrismaClient) {
             author: {
                 select: {
                     id: true,
-                    email: true
-                }
+                    email: true,
+                },
             },
             tags: true,
             images: true,
             Comments: true,
             category: true,
-
-        }
-    })
+        },
+    });
     if (existingblog) return existingblog;
-
 }
 
-async function getAll(page: number, pageSize: number, prismaClient: PrismaClient) {
+async function getAll(
+    page: number,
+    pageSize: number,
+    prismaClient: PrismaClient,
+) {
     const blogs = prismaClient.blog;
 
-    if (pageSize !== 10 && pageSize != 30 && pageSize !== 50 && pageSize !== 0) throw new Error('page size must be 10, 30 or 50')
+    if (pageSize !== 10 && pageSize != 30 && pageSize !== 50 && pageSize !== 0)
+        throw new Error("page size must be 10, 30 or 50");
 
     let allBlogs = await blogs.findMany({
-        skip: page === 0 ? 0 : (page - 1) * pageSize, take: page === 0 ? 9999 : pageSize,
+        skip: page === 0 ? 0 : (page - 1) * pageSize,
+        take: page === 0 ? 9999 : pageSize,
         where: {
             publishDate: {
-                lte: new Date()
-            }
+                lte: new Date(),
+            },
         },
         include: {
             // reviews: true,
             tags: true,
             author: true,
             category: true,
-
-        }
-    })
+        },
+    });
 
     const totalCount = await blogs.count();
     const totalPages = Math.ceil(totalCount / pageSize);
 
-    return { records: allBlogs, currentPage: page, totalPages, pageSize }
-
+    return { records: allBlogs, currentPage: page, totalPages, pageSize };
 }
 
 export async function getFeatured(prisma: PrismaClient) {
@@ -116,38 +121,78 @@ export async function getFeatured(prisma: PrismaClient) {
         where: {
             featured: true,
             publishDate: {
-                lte: new Date()
-            }
+                lte: new Date(),
+            },
         },
         orderBy: {
-            publishDate: 'desc'
+            publishDate: "desc",
         },
         include: {
             tags: true,
             author: {
                 include: {
                     image: true,
-                }
+                },
             },
             images: true,
-            category: true
-        }
+            category: true,
+        },
     });
     return getRandomFromArray(featured);
+}
+
+export async function getBlogsByCategory(id: string, prisma: PrismaClient) {
+
+    const list = await prisma.blog.findMany({
+        where: {
+            OR: [
+                {
+                    category: {
+                        parent: {
+                            id: id
+                        }
+                    }
+                }, {
+                    category: {
+                        id: id
+                    }
+                }
+
+            ],
+
+        },
+        include: {
+            tags: true,
+            author: {
+                include: {
+                    image: true,
+                },
+            },
+            images: true,
+            category: true,
+        },
+    });
+    return list
 
 }
-export async function addView({ id, userEmail }: { id: string, userEmail?: string }, prisma: PrismaClient) {
+
+
+
+export async function addView(
+    { id, userEmail }: { id: string; userEmail?: string },
+    prisma: PrismaClient,
+) {
     const blogs = prisma.blog;
     try {
         const update = await blogs.update({
             where: {
                 id,
                 publishDate: {
-                    lte: new Date()
-                }
+                    lte: new Date(),
+                },
             },
             data: {
-                Views: { increment: 1 }
+                Views: { increment: 1 },
             },
             include: {
                 tags: true,
@@ -156,79 +201,80 @@ export async function addView({ id, userEmail }: { id: string, userEmail?: strin
                 author: {
                     include: {
                         image: true,
-                    }
+                    },
                 },
                 images: true,
-                Likes: userEmail ? {
-                    where: {
-                        user: {
-                            email: userEmail
-                        }
+                Likes: userEmail
+                    ? {
+                        where: {
+                            user: {
+                                email: userEmail,
+                            },
+                        },
                     }
-                } : false,
+                    : false,
                 _count: {
                     select: {
-                        Likes: true
-                    }
+                        Likes: true,
+                    },
                 },
                 Comments: {
                     include: {
-                        User: true
-                    }
+                        User: true,
+                    },
                 },
-            }
-        })
-    
-        return update
+            },
+        });
+
+        return update;
     } catch (error) {
         console.log(error);
-        return 
-        
+        return;
     }
 }
 export function getRecent(prisma: PrismaClient) {
-    const recentDate = new Date(Date.now() - 90 * (24 * 60 * 60 * 1000)) // 90 days
+    const recentDate = new Date(Date.now() - 90 * (24 * 60 * 60 * 1000)); // 90 days
     const recent = prisma.blog.findMany({
         where: {
             AND: [
                 {
                     publishDate: {
-                        lte: new Date()
-                    }
+                        lte: new Date(),
+                    },
                 },
                 {
                     publishDate: {
-                        gte: recentDate
-                    }
+                        gte: recentDate,
+                    },
                 },
-            ]
-
+            ],
         },
         include: {
             tags: true,
             category: true,
 
             author: {
-
                 include: {
                     image: true,
-                }
-            }, images: true
+                },
+            },
+            images: true,
         },
         orderBy: {
-            date: 'desc'
-        }
+            date: "desc",
+        },
     });
     return recent;
 }
 
 export function getPopular(prisma: PrismaClient) {
     const popular = prisma.blog.findMany({
-        skip: 0, take: 10,
+        skip: 0,
+        take: 10,
         where: {
             publishDate: {
-                lte: new Date()
-            }
+                lte: new Date(),
+            },
         },
         include: {
             // reviews: true,
@@ -236,31 +282,29 @@ export function getPopular(prisma: PrismaClient) {
             category: true,
 
             author: {
-
                 include: {
                     image: true,
-                }
-            }, images: true
-
-
+                },
+            },
+            images: true,
         },
         orderBy: {
             Likes: {
-                _count: 'asc'
+                _count: "asc",
             },
-        }
-    })
-    return popular
+        },
+    });
+    return popular;
 }
 
 export function getEssential(prisma: PrismaClient) {
-
     const essential = prisma.blog.findMany({
-        skip: 0, take: 10,
+        skip: 0,
+        take: 10,
         where: {
             publishDate: {
-                lte: new Date()
-            }
+                lte: new Date(),
+            },
         },
         include: {
             // reviews: true,
@@ -269,39 +313,34 @@ export function getEssential(prisma: PrismaClient) {
 
             author: {
                 include: {
-                    image: true
-                }
+                    image: true,
+                },
             },
-            images: true
-
-
-
-        }
-    })
-    return essential
+            images: true,
+        },
+    });
+    return essential;
 }
 
 async function getAuthor(id: string, page: number, prisma: PrismaClient) {
     const users = prisma.user;
 
     const author = await users.findUnique({
-        where: { id }, include: {
-
+        where: { id },
+        include: {
             image: true,
             blogs: {
                 take: 10,
                 skip: (page - 1) * 10,
                 include: {
-                    images: true
-                }
-            }
-        }
-    })
+                    images: true,
+                },
+            },
+        },
+    });
 
-    return author
-
+    return author;
 }
-
 
 async function addComment(comment: CommentDTO, prisma: PrismaClient) {
     const comments = prisma.blogComment;
@@ -311,33 +350,104 @@ async function addComment(comment: CommentDTO, prisma: PrismaClient) {
             comment: comment.comment,
             User: {
                 connect: {
-                    email: comment.email
-                }
+                    email: comment.email,
+                },
             },
             Blog: { connect: { id: comment.blogId } },
         },
         include: {
-            User: true
-        }
-    })
+            User: true,
+        },
+    });
 
-    return newComment
+    return newComment;
 }
 
 async function getComments(id: string, page: number, prisma: PrismaClient) {
     const comments = prisma.blogComment;
 
     const readComments = await comments.findMany({
-        skip: (page - 1) * 5, take: 5,
+        skip: (page - 1) * 5,
+        take: 5,
         where: {
             blogId: id,
-        }
-    })
+        },
+    });
 
-    return readComments
+    return readComments;
 }
 
-export async function getBySearchTerm(search: string, page: number, prisma: PrismaClient) {
+export async function getSimilar(id: string, prisma: PrismaClient) {
+    const blog = await prisma.blog.findUnique({
+        where: {
+            id,
+        },
+        include: {
+            category: {
+                include: {
+                    parent: true,
+                },
+            },
+            tags: true,
+        },
+    });
+
+    if (!blog) return [];
+    else {
+        const similar = await prisma.blog.findMany({
+            skip: 0,
+            take: 10,
+            where: {
+                OR: [
+                    {
+                        category: {
+                            id: {
+                                equals: blog.category?.id ?? "",
+                            },
+                        },
+                    },
+
+                    {
+                        category: {
+                            parentId: {
+                                equals: blog.category?.parent?.id ?? "",
+                            },
+                        },
+                    },
+
+                    {
+                        tags: {
+                            every: {
+                                name: {
+                                    in: blog.tags.map((t) => t.name),
+                                },
+                            },
+                        },
+                    },
+                ],
+            },
+            include: {
+                tags: true,
+                category: true,
+
+                author: {
+                    include: {
+                        image: true,
+                    },
+                },
+                images: true,
+            },
+        });
+
+
+        return similar as DisplayBlogDTO[];
+    }
+}
+export async function getBySearchTerm(
+    search: string,
+    page: number,
+    prisma: PrismaClient,
+) {
     const blogs = prisma.blog;
     const records = await blogs.findMany({
         where: {
@@ -346,88 +456,100 @@ export async function getBySearchTerm(search: string, page: number, prisma: Pris
                     title: {
                         contains: search,
                     },
-
                 },
                 {
                     description: {
                         contains: search,
-                    }
+                    },
                 },
                 {
                     subTitle: {
-                        contains: search
-                    }
+                        contains: search,
+                    },
                 },
                 {
                     content: {
                         contains: cleanHtmlString(search),
-                    }
-                }
+                    },
+                },
             ],
-            AND: [{
-                publishDate: {
-                    lte: new Date()
-                }
-            }]
+            AND: [
+                {
+                    publishDate: {
+                        lte: new Date(),
+                    },
+                },
+            ],
         },
         orderBy: {
-            publishDate: 'desc'
-        }
-    })
-    return records
-
+            publishDate: "desc",
+        },
+    });
+    return records;
 }
 
-export async function addLike(blogId: string, userEmail: string, prisma: PrismaClient) {
+export async function addLike(
+    blogId: string,
+    userEmail: string,
+    prisma: PrismaClient,
+) {
     const BlogLike = prisma.blogLike;
     const Like = await BlogLike.create({
         data: {
             blog: {
-                connect: { id: blogId }
+                connect: { id: blogId },
             },
             user: {
-                connect: { email: userEmail }
-            }
-        }
-    })
+                connect: { email: userEmail },
+            },
+        },
+    });
 
     const newLikes = await BlogLike.count({
         where: {
             blog: {
-                id: blogId
-            }
-        }
-    })
-    return { liked: true, likes: newLikes }
-
-
+                id: blogId,
+            },
+        },
+    });
+    return { liked: true, likes: newLikes };
 }
 
-export async function removeLike(blogId: string, userEmail: string, prisma: PrismaClient) {
+export async function removeLike(
+    blogId: string,
+    userEmail: string,
+    prisma: PrismaClient,
+) {
     const BlogLike = prisma.blogLike;
     const user = await getUserByEmail(userEmail, prisma);
-    if (!user) return false
+    if (!user) return false;
     const Like = await BlogLike.delete({
         where: {
             userId_blogId: {
                 blogId,
-                userId: user.id
+                userId: user.id,
             },
-
-
-        }
-    })
+        },
+    });
 
     const newLikes = await BlogLike.count({
         where: {
             blog: {
-                id: blogId
-            }
-        }
-    })
+                id: blogId,
+            },
+        },
+    });
 
-    return { liked: false, likes: newLikes }
-
+    return { liked: false, likes: newLikes };
 }
 
-export { create, update, remove, read, getAll, getAuthor, addComment, getComments }
+export {
+    create,
+    update,
+    remove,
+    read,
+    getAll,
+    getAuthor,
+    addComment,
+    getComments,
+};
