@@ -224,7 +224,8 @@ async function getAll(page: number, pageSize: number, prismaClient: PrismaClient
         include: {
             image: true,
             tags: true,
-            SubServices: true
+            SubServices: true,
+            CaseStudies:  true
         },
         orderBy: {
 
